stack: use byte array for channel masks
The RCP API has been maintained as it was. In fact using
spinel_push_fixed_u8_array results in the same byte order as using
spinel_push_fixed_u32_array.

A 32bit unsigned integer b31..b0 is written in little endian in spinel:

b7..b0 b15..b8 b23..b16 b31..b24

Which is the same as 4 8bit unsigned integers in order.

WI_SUN-1374

+++ app_wshwsim/fhss/channel_list.h
@@ -24,6 +24,6 @@
  * @param list to scan
  * @return amount of bits set in the channel masks
  */
-int channel_list_count_channels(const uint32_t *list);
+int channel_list_count_channels(const uint8_t *list);
 
 #endif // !__CHANNEL_LIST_H__
