+++ stack/source/6lowpan/ws/ws_bootstrap.h
@@ -176,7 +176,7 @@ void ws_bootstrap_configure_csma_ca_backoffs(struct net_if *cur, uint8_t max_bac
 void ws_bootstrap_fhss_configure_channel_masks(struct net_if *cur, fhss_ws_configuration_t *fhss_configuration);
 int8_t ws_bootstrap_fhss_set_defaults(struct net_if *cur, fhss_ws_configuration_t *fhss_configuration);
 void ws_bootstrap_fhss_activate(struct net_if *cur);
-uint16_t ws_bootstrap_randomize_fixed_channel(uint16_t configured_fixed_channel, uint8_t number_of_channels, uint32_t *channel_mask);
+uint16_t ws_bootstrap_randomize_fixed_channel(uint16_t configured_fixed_channel, uint8_t number_of_channels, uint8_t *channel_mask);
 int ws_bootstrap_set_domain_rf_config(struct net_if *cur);
 void ws_bootstrap_configure_max_retries(struct net_if *cur, uint8_t max_mac_retries);
 void ws_bootstrap_configure_data_request_restart(struct net_if *cur, uint8_t cca_failure_restart_max, uint8_t tx_failure_restart_max, uint16_t blacklist_min_ms, uint16_t blacklist_max_ms);
